device scanner implemented

+++ internal/models/BLEdevice.go
@@ -4,5 +4,5 @@ package models
 type BLEDevice struct {
 	Name    string `json:"name"`    // The name of the BLE device (e.g., "Smart Scale")
 	Address string `json:"address"` // The MAC address of the BLE device
-	RSSI    int    `json:"rssi"`    // Received Signal Strength Indicator (RSSI) in dBm
+	RSSI    int16  `json:"rssi"`    // Received Signal Strength Indicator (RSSI) in dBm
 }
